DTSBPS-698: move final package name util class with filename helper
